refactor(adf): split bundler into manifest.ts and merger.ts (closes #8)
Extract manifest parsing + trigger resolution into manifest.ts and pure
merge logic + token estimation into merger.ts. bundler.ts remains as a
thin orchestration shell with re-exports for backward compatibility.

All 233 tests pass (215 existing unchanged + 18 new focused tests).
